requirements.txt
playwright==1.58.0
pytest==9.0.2
pytest-playwright==0.7.2
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-rerunfailures==16.1
python-dotenv==1.2.2

# Mobile (Appium) — selenium is included as a transitive dependency
Appium-Python-Client
